+++ app/Http/Controllers/OrderController.php
@@ -19,12 +19,15 @@ class OrderController extends Controller
 
     public function index()
     {
+        $this->authorize('viewIndex', Order::class);
         return Order::with('menu_items')->get();
     }
 
     public function show($id)
     {
-        return Order::find($id)->with('menu_items')->get();
+        $order = Order::findOrFail($id)->with('menu_items')->first();
+        $this->authorize('view', $order);
+        return $order;
     }
 
     public function store(Request $request)
@@ -37,6 +40,7 @@ class OrderController extends Controller
     public function update(Request $request, $id)
     {
         $order = Order::findOrFail($id);
+        $this->authorize('update', $order);
         $order->update($request->all());
 
         return response()->json($order, 200);
@@ -45,6 +49,7 @@ class OrderController extends Controller
     public function destroy(Request $request, $id)
     {
         $order = Order::findOrFail($id);
+        $this->authorize('delete', $order);
         $order->delete();
 
         return response()->json(null, 204);
